pyproject.toml
[project]
name = "rscience-causal"
version = "0.1.0"
description = "Causal relationship extraction and visualization from r/science submissions"
requires-python = ">=3.12"

dependencies = [
    # Data access
    "duckdb>=1.1.0",
    "pyarrow>=18.0.0",

    # Pipeline core
    "pyyaml>=6.0",
    "pydantic>=2.7.0",

    # Step 2 default
    "spacy>=3.7.0",
    # Run after install: python -m spacy download en_core_web_sm

    # Step 3 default
    "sentence-transformers>=3.0.0",
    "hdbscan>=0.8.38",
    "scikit-learn>=1.5.0",
    "numpy>=1.26.0",

    # API
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
llm = [
    "openai>=1.30.0",
    "anthropic>=0.28.0",
]
zero-shot = [
    "transformers>=4.40.0",
    "torch>=2.3.0",
]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["pipeline", "api"]

[tool.ruff]
line-length = 100
target-version = "py312"
